add 'Import from a Directory' operation
