+++ setup.cfg
@@ -1,8 +1,7 @@
 [flake8]
 per-file-ignores =
     addons/blender_dds_addon/__init__.py:F821,F722
-    addons/blender_dds_addon/import_dds.py:F821,F722
-    addons/blender_dds_addon/export_dds.py:F821,F722
+    addons/blender_dds_addon/ui/*.py:F821,F722
     tests/*:F821,F722
 
 max-line-length = 119
