+++ tests/test_dds.py
@@ -4,11 +4,11 @@ import pytest
 
 from blender_dds_addon.ui import import_dds
 from blender_dds_addon.ui import export_dds
-from blender_dds_addon.ui import set_properties
+from blender_dds_addon.ui import custom_properties
 
 import bpy
-bpy.utils.register_class(set_properties.DDSProperties)
-set_properties.add_custom_props_for_dds()
+bpy.utils.register_class(custom_properties.DDSCustomProperties)
+custom_properties.add_custom_props_for_dds()
 
 
 def get_test_dds():
